chore: changes in migration

+++ src/adminPanel/infrastructure/repository/sequelize/migrations/20240213041319-report-agency-properties-table.js
@@ -33,7 +33,7 @@ async function up(queryInterface, Sequelize) {
       allowNull: false
     },
     scoredProbabilities: {
-      type: Sequelize.INTEGER,
+      type: Sequelize.DOUBLE(10, 2),
       allowNull: false
     },
     firstPredictedDate: {
